Conf tests are passing
