+++ .github/scripts/test-info.mjs
@@ -623,6 +623,7 @@ const components = {
             'state/cockroachdb',
             'common/component/postgresql',
             'common/component/sql',
+            'common/component/sql/migrations',
         ],
     },
     'state.etcd.v1': {
@@ -664,23 +665,29 @@ const components = {
         conformance: true,
         conformanceSetup: 'docker-compose.sh oracledatabase',
     },
-    'state.postgresql': {
+    'state.postgresql.v1': {
         certification: true,
         sourcePkg: [
-            'state/postgresql',
+            'state/postgresql/v1',
             'common/authentication/postgresql',
-            'common/component/postgresql',
+            'common/component/postgresql/interfaces',
+            'common/component/postgresql/transactions',
+            'common/component/postgresql/v1',
             'common/component/sql',
+            'common/component/sql/migrations',
         ],
     },
     'state.postgresql.v1.docker': {
         conformance: true,
         conformanceSetup: 'docker-compose.sh postgresql',
         sourcePkg: [
-            'state/postgresql',
+            'state/postgresql/v1',
             'common/authentication/postgresql',
-            'common/component/postgresql',
+            'common/component/postgresql/interfaces',
+            'common/component/postgresql/transactions',
+            'common/component/postgresql/v1',
             'common/component/sql',
+            'common/component/sql/migrations',
         ],
     },
     'state.postgresql.v1.azure': {
@@ -692,10 +699,42 @@ const components = {
             'AzureDBPostgresTenantId',
         ],
         sourcePkg: [
-            'state/postgresql',
+            'state/postgresql/v1',
             'common/authentication/postgresql',
-            'common/component/postgresql',
+            'common/component/postgresql/interfaces',
+            'common/component/postgresql/transactions',
+            'common/component/postgresql/v1',
             'common/component/sql',
+            'common/component/sql/migrations',
+        ],
+    },
+    'state.postgresql.v2.docker': {
+        conformance: true,
+        conformanceSetup: 'docker-compose.sh postgresql',
+        sourcePkg: [
+            'state/postgresql/v2',
+            'common/authentication/postgresql',
+            'common/component/postgresql/interfaces',
+            'common/component/postgresql/transactions',
+            'common/component/sql',
+            'common/component/sql/migrations',
+        ],
+    },
+    'state.postgresql.v2.azure': {
+        conformance: true,
+        requiredSecrets: [
+            'AzureDBPostgresConnectionString',
+            'AzureDBPostgresClientId',
+            'AzureDBPostgresClientSecret',
+            'AzureDBPostgresTenantId',
+        ],
+        sourcePkg: [
+            'state/postgresql/v2',
+            'common/authentication/postgresql',
+            'common/component/postgresql/interfaces',
+            'common/component/postgresql/transactions',
+            'common/component/sql',
+            'common/component/sql/migrations',
         ],
     },
     'state.redis': {
