+++ tests/certification/state/postgresql/v1/components/docker/default/postgresqlstatestore.yaml
