+++ tests/config/state/tests.yml
@@ -44,6 +44,16 @@ components:
     config:
       # This component requires etags to be numeric
       badEtag: "1"
+  - component: postgresql.v2.docker
+    operations: [ "transaction", "etag", "first-write", "ttl" ]
+    config:
+      # This component requires etags to be UUIDs
+      badEtag: "e9b9e142-74b1-4a2e-8e90-3f4ffeea2e70"
+  - component: postgresql.v2.azure
+    operations: [ "transaction", "etag", "first-write", "ttl" ]
+    config:
+      # This component requires etags to be UUIDs
+      badEtag: "e9b9e142-74b1-4a2e-8e90-3f4ffeea2e70"
   - component: sqlite
     operations: [ "transaction", "etag",  "first-write", "ttl" ]
   - component: mysql.mysql
